- WorldToScreen signature!
At last Morphling reversed the process and now we can rest easy
+ AbilityESP rounding replaced by percentage

+++ Dota2Cheat/CheatSDK/Config.h
@@ -77,7 +77,7 @@ namespace Config {
 		inline bool Enabled;
 		inline bool ShowAllies;
 		inline float UIScale;
-		inline float Rounding;
+		inline int Rounding;
 		inline int LevelCounterType;
 		inline bool ShowCooldownDecimals;
 	}
